backend/requirements.txt
django==4.2.16  # Adjust the version as needed
google-generativeai==0.8.2  # Include the package you're using for Google AI
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-environ==0.11.2
mem0ai==0.1.19
litellm==1.48.14
